+++ resources/assets/components/pages/AccountPage/AccountContainer.js
@@ -1,10 +1,10 @@
 import { connect } from 'react-redux';
 
-import ProfileQuery from './ProfileQuery';
+import AccountQuery from './AccountQuery';
 import { getUserId } from '../../../selectors/user';
 
 const mapStateToProps = state => ({
   userId: getUserId(state),
 });
 
-export default connect(mapStateToProps)(ProfileQuery);
+export default connect(mapStateToProps)(AccountQuery);
